Sat, Apr 27, 2024, 11:30 PM +02:00

+++ caesar/caesar.c
@@ -54,7 +54,7 @@ char rotate(char p, int k)
     printf("The char %c is %i \n", p, p);
     if(isalpha(p))
     {
-        c = (p + k) % 26;
+        c = ((int) p + k) % 26;
         return (char) c;
     }
     else
